# Compiler perf: Count line numbers progressively