This is another attempt to squeeze perf out of `compileAsync`. Previous attempt here: https://github.com/microsoft/pxt/pull/7283, which brought it from 20s to 6s. That was a nice improvement, but still noticeably slow. Getting better results this time:

Time spent in `compileAsync`:
| Before | After |
|---|---|
| 6s | 500ms |

Testing with this project: https://arcade.makecode.com/74548-67333-76127-33508

Related: https://github.com/microsoft/pxt-arcade/issues/1928