@eanders-MS 2020-07-09
Hm, the PR title isn't isn't accurate (will update). Still using a regexp, but no longer scanning the whole program each time.